feat: add auth

+++ src/modules/tasks/tasks.service.ts
@@ -2,22 +2,34 @@ import { prisma } from '../../database/prisma.js';
 
 import { createTaskNotFoundError } from '../../errors/task-not-found-error.js';
 
-export async function createTask(title: string) {
+export async function createTask(
+  userId: number,
+  title: string,
+) {
   return prisma.task.create({
     data: {
       title,
+      userId,
     },
   });
 }
 
-export async function findAllTasks() {
-  return prisma.task.findMany();
+export async function findAllTasks(userId: number) {
+  return prisma.task.findMany({
+    where: {
+      userId,
+    },
+  });
 }
 
-export async function findTaskById(id: number) {
-  const task = await prisma.task.findUnique({
+export async function findTaskById(
+  userId: number,
+  id: number,
+) {
+  const task = await prisma.task.findFirst({
     where: {
       id,
+      userId,
     },
   });
 
@@ -29,12 +41,24 @@ export async function findTaskById(id: number) {
 }
 
 export async function updateTask(
+  userId: number,
   id: number,
   data: {
     title?: string;
     completed?: boolean;
   },
 ) {
+  const task = await prisma.task.findFirst({
+    where: {
+      id,
+      userId,
+    },
+  });
+
+  if (!task) {
+    throw createTaskNotFoundError();
+  }
+
   return prisma.task.update({
     where: {
       id,
@@ -43,7 +67,21 @@ export async function updateTask(
   });
 }
 
-export async function deleteTask(id: number) {
+export async function deleteTask(
+  userId: number,
+  id: number,
+) {
+  const task = await prisma.task.findFirst({
+    where: {
+      id,
+      userId,
+    },
+  });
+
+  if (!task) {
+    throw createTaskNotFoundError();
+  }
+
   return prisma.task.delete({
     where: {
       id,
